# Speedup element equality checks in ConfigurationPropertyName