Hi,

here is another PR that tries to improve the binding of extremely large files as reported in #16401. 

There are basically two main ideas behind the proposed changes:

1. Change `isAncestorOf()` so it is more likely inlined.
2. When profiling the provided app in #16401 it became apparent that the element equality checks compare mostly a combination of `UNIFORM`, `NUMERICALLY_INDEXED` and `DASHED` elements. All of those `ElementType`s don't really need any lowercasing or skipping of non alphanumeric chars. Specifically, the first two could benefit from a fast version. `DASHED` - described as "almost" uniform - can benefit of a trimmed down version of the current checks as well.

The proposed changes provide the following results with @wilkinsona`s benchmark harness:

|       | Baseline |  New  |
| ----- | -------: | ----: |
|       |    15.109 | 12.921 |
|       |    13.843 | 12.042 |
|       |    13.451 | 11.900 |
|       |    12.736 | 11.975 |
|       |    12.712 | 11.945 |
|       |    14.475 | 12.289 |
|       |    13.132 | 11.821 |
|       |    13.066 | 12.177 |
|       |    12.261 | 12.343 |
|       |    12.940 | 12.048 |
|  Mean |    13.372 | 12.146 |
| Range |    2.848 | 1.100 |

When running an adjusted version of the JMH benchmarks provided in #15760 I get the following results:
**Before**
```
Benchmark                           (prof)   Mode  Cnt      Score      Error  Units
PropertiesBenchmarkIT.auto          medium  thrpt   10    194,016 ±  107,644  ops/s
PropertiesBenchmarkIT.auto:size     medium  thrpt   10   1340,000                 #
PropertiesBenchmarkIT.auto           small  thrpt   10  13102,403 ± 6903,192  ops/s
PropertiesBenchmarkIT.auto:size      small  thrpt   10     10,000                 #
PropertiesBenchmarkIT.auto           large  thrpt   10      2,886 ±    0,980  ops/s
PropertiesBenchmarkIT.auto:size      large  thrpt   10  10930,000                 #
PropertiesBenchmarkIT.auto       verylarge  thrpt   10      0,119 ±    0,016  ops/s
PropertiesBenchmarkIT.auto:size  verylarge  thrpt   10  50000,000                 #
```

**After**
```
Benchmark                           (prof)   Mode  Cnt      Score      Error  Units
PropertiesBenchmarkIT.auto          medium  thrpt   10    203,827 ±  104,316  ops/s
PropertiesBenchmarkIT.auto:size     medium  thrpt   10   1340,000                 #
PropertiesBenchmarkIT.auto           small  thrpt   10  15183,887 ± 5609,828  ops/s
PropertiesBenchmarkIT.auto:size      small  thrpt   10     10,000                 #
PropertiesBenchmarkIT.auto           large  thrpt   10      3,015 ±    0,898  ops/s
PropertiesBenchmarkIT.auto:size      large  thrpt   10  10930,000                 #
PropertiesBenchmarkIT.auto       verylarge  thrpt   10      0,147 ±    0,009  ops/s
PropertiesBenchmarkIT.auto:size  verylarge  thrpt   10  50000,000                 #
```

Overall, I think these are promising results. Let me know what you think.

Cheers,
Christoph