@wilkinsona 2019-04-24
@dreis2211 Thanks very much for this. I've had the time (and brain power) to hopefully give things a proper review and the changes seem good to me. I spent some time wondering if we needed some more tests, but things already seem to be covered quite nicely by the existing `ConfigurationPropertyNameTests.equalsAndHashCode()`.