@dreis2211 2019-04-24
Thanks for the review. `ConfigurationPropertyNameTests.equalsAndHashCode()` did indeed fail while I worked on this, so I figured the same.