@Shawyeok 2021-06-30
@wilkinsona @dreis2211 
Nice work!
I'm face this problem too, but currently my project still use spring-boot `2.1.x`, is `2.1.x` under maintenance? Can this improvement pick to `2.1.x` branch?