@wilkinsona 2021-06-30
I'm afraid not, no. As [announced at the end of 2019](https://spring.io/blog/2019/12/10/spring-boot-2-1-x-eol-november-1st-2020), Spring Boot 2.1.x reached the end of its supported life in November 2020. You should upgrade to Spring Boot 2.4.x or 2.5.x at your earliest convenience.